+++ src/StateR/AsyncLogic/AsyncOperation.cs
@@ -46,7 +46,7 @@ namespace StateR.AsyncLogic
         protected virtual async Task HandleExceptionAsync(IDispatchContext<TAction> context, TState state, Exception ex, CancellationToken cancellationToken)
         {
             await DispatchStatusUpdateAsync(AsyncOperationStatus.Failed, cancellationToken);
-            var errorAction = new AsyncErrorOccured(context.Action, state, Store.GetState<TState>(), ex);
+            var errorAction = new AsyncError.Occured(context.Action, state, Store.GetState<TState>(), ex);
             await Store.DispatchAsync(errorAction, cancellationToken);
         }
 
@@ -56,79 +56,5 @@ namespace StateR.AsyncLogic
         }
 
         protected abstract Task<TSuccessAction> LoadAsync(TAction action, TState initalState, CancellationToken cancellationToken = default);
-    }
-    public abstract record AsyncState : StateBase
-    {
-        public AsyncOperationStatus Status { get; init; }
-    }
-    public enum AsyncOperationStatus
-    {
-        Idle,
-        Loading,
-        Succeeded,
-        Failed,
-    }
-    public record StatusUpdated(AsyncOperationStatus status) : IAction;
-
-    // AsyncErrorOccured
-    public record AsyncErrorOccured(IAction Action, AsyncState InitialState, AsyncState ActualState, Exception Exception) : IAction;
-    public record AsyncErrorState : StateBase
-    {
-        public IAction Action { get; init; }
-        public AsyncState InitialState { get; init; }
-        public AsyncState ActualState { get; init; }
-        public Exception Exception { get; init; }
-
-        public bool HasException() => Exception != null;
-        public bool HasActualState() => ActualState != null;
-        public bool HasInitialState() => InitialState != null;
-        public bool HasAction() => Action != null;
-    }
-
-    public class InitialAsyncErrorState : IInitialState<AsyncErrorState>
-    {
-        public AsyncErrorState Value => new();
-    }
-
-    public class AsyncErrorOccuredReducer : IReducer<AsyncErrorOccured, AsyncErrorState>
-    {
-        public AsyncErrorState Reduce(AsyncErrorOccured action, AsyncErrorState initialState) => initialState with {
-            Action = action.Action,
-            InitialState = action.InitialState,
-            ActualState = action.ActualState,
-            Exception = action.Exception
-        };
-    }
-
-    public static class AsyncLogicStartupExtensions
-    {
-        public static IStatorBuilder AddAsyncOperations(this IStatorBuilder builder)
-        {
-            builder.AddTypes(new[] {
-                typeof(StatusUpdated),
-            });
-            return builder.AddAsyncErrors();
-        }
-
-        public static IStatorBuilder AddAsyncErrors(this IStatorBuilder builder)
-        {
-            ////var asyncStateType = typeof(AsyncState);
-            //var types = new[]
-            //{
-            //    typeof(ReducerHandler<AsyncErrorState, AsyncErrorOccured>),
-            //    typeof(AsyncErrorOccuredReducer),
-            //    typeof(InitialAsyncErrorState),
-            //    typeof(AsyncErrorState)
-            //};
-            ////var types = asyncStateType.Assembly.GetTypes().Where(t => t.Namespace == asyncStateType.Namespace);
-            //builder.AddTypes(types);
-            builder.Services
-                .AddSingleton<IActionHandler<AsyncErrorOccured>, ReducerHandler<AsyncErrorState, AsyncErrorOccured>>()
-                .AddSingleton<IReducer<AsyncErrorOccured, AsyncErrorState>, AsyncErrorOccuredReducer>()
-                .AddSingleton<IInitialState<AsyncErrorState>, InitialAsyncErrorState>()
-                .AddSingleton<IState<AsyncErrorState>, Internal.State<AsyncErrorState>>()
-            ;
-            return builder;
-        }
     }
 }
