+++ src/StateR/AsyncLogic/AsyncOperationStatus.cs
@@ -1,0 +1,10 @@
+﻿namespace StateR.AsyncLogic
+{
+    public enum AsyncOperationStatus
+    {
+        Idle,
+        Loading,
+        Succeeded,
+        Failed,
+    }
+}
