+++ src/StateR/AsyncLogic/AsyncState.cs
@@ -1,0 +1,7 @@
+﻿namespace StateR.AsyncLogic
+{
+    public abstract record AsyncState : StateBase
+    {
+        public AsyncOperationStatus Status { get; init; }
+    }
+}
