+++ src/StateR/AsyncLogic/StatusUpdated.cs
@@ -1,0 +1,4 @@
+﻿namespace StateR.AsyncLogic
+{
+    public record StatusUpdated(AsyncOperationStatus status) : IAction;
+}
